Remove duplicate app_main(). Added ap-test to list of tests. Renamed AP test.
