+++ CMakeLists.txt
@@ -20,7 +20,7 @@ cmake_minimum_required(VERSION 3.10)
 set(CMAKE_CXX_STANDARD 17)
 
 # Select the test project to build
-set(selected_test_project sntp)
+set(selected_test_project access_point)
 
 # For Linux builds, you may enable address sanitizer
 set(SMOOTH_ENABLE_ASAN 0)
@@ -29,6 +29,7 @@ set(SMOOTH_ASAN_OPTIMIZATION_LEVEL 1)
 list(APPEND available_tests
         hello_world
         asan_test
+        access_point
         logging
         json_test
         mqtt
